python/requirements.txt
mysql-connector-python>=9.0.0
